ml/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
Flask-Cors==4.0.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.8.2
pydantic-core==2.20.1
sniffio==1.3.1
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.20.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0